+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/headless_in_app_webview/HeadlessInAppWebViewManager.kt
@@ -53,7 +53,7 @@ open class HeadlessInAppWebViewManager(
             context,
             id,
             params,
-            deferNativeOperationsUntilAttach
+            false
         )
         val headlessInAppWebView = HeadlessInAppWebView(currentPlugin, id, flutterWebView)
         webViews[id] = headlessInAppWebView