+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/webview/in_app_webview/InAppWebViewChromeClient.kt
@@ -148,6 +148,7 @@ open class InAppWebViewChromeClient(
     inAppWebView?.let { webView ->
       webView.channelDelegate?.onExitFullscreen()
       webView.setInFullscreen(false)
+      webView.restoreInputConnectionAfterFullscreen()
     }
   }